perf: ini og screen
